remove draggablelist component, rename discussion timer component
